Split common and configuration templates
